Moved print, println and readline functions to Pure namespace
