+++ StepLang/Framework/Pure/PrintFunction.cs
@@ -3,7 +3,7 @@ using StepLang.Framework.Conversion;
 using StepLang.Interpreting;
 using StepLang.Parsing.Expressions;
 
-namespace StepLang.Framework.IO;
+namespace StepLang.Framework.Pure;
 
 public class PrintFunction : NativeFunction
 {
