+++ StepLang/Interpreting/Scope.cs
@@ -1,6 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
 using StepLang.Framework.Conversion;
-using StepLang.Framework.IO;
 using StepLang.Framework.Mutating;
 using StepLang.Framework.Other;
 using StepLang.Framework.Pure;
